test: 기도 테스트 공통 셋업 함수 분리 #8
