+++ internal/room/create_room_api_test.go
@@ -11,37 +11,7 @@ import (
 	sharedHttp "github.com/changhyeonkim/pray-together/go-api-server/internal/shared/http"
 	"github.com/changhyeonkim/pray-together/go-api-server/internal/shared/testutil"
 	"github.com/stretchr/testify/assert"
-	"gorm.io/gorm"
 )
-
-// setupTestEnvironment creates all dependencies needed for room handler tests
-// Returns the handler, database, and a test member
-func setupTestEnvironment(t *testing.T) (*room.RoomHandler, *gorm.DB, int64) {
-	t.Helper()
-
-	// Setup test database
-	db := testutil.SetupTestDB(t)
-	t.Cleanup(func() {
-		testutil.CleanupTestDB(t, db)
-	})
-
-	// Create a test member for authenticated requests
-	testMember := testutil.CreateTestMember(t, db)
-
-	// Setup dependencies - need to import member package
-	roomRepo := room.NewRoomRepository()
-	memberRoomRepo := room.NewMemberRoomRepository()
-
-	// Create member service for validation
-	memberRepo := testutil.NewMemberRepository(db)
-	memberService := testutil.NewMemberService(memberRepo)
-
-	roomService := room.NewRoomService(roomRepo, memberRoomRepo, memberService)
-	roomUseCase := room.NewRoomUseCase(db, roomService)
-	roomHandler := room.NewRoomHandler(roomUseCase)
-
-	return roomHandler, db, int64(testMember.ID)
-}
 
 func TestCreateRoom_Success(t *testing.T) {
 	// Given: Setup test environment
